requirements.txt
google-generativeai
pypdf
python-dotenv
openai
streamlit
fpdf2
